requirements.txt
pandas==1.4.3
beautifulsoup4==4.11.1
matplotlib== 3.9.4
transbigdata==0.5.3
geopandas==0.6.1
PyYAML==6.0
lxml==4.9.2
fake-useragent==2.2.0